implement 'bleeding-edge' dependencies
With "*" as a dependency's version, if its repository exists on disk when
running 'gocart', it will be left alone (not locked down). Its current version
will be saved to the .lock.

If the repository does NOT exist, it will be initialized and locked down to
the version in the .lock. This is the case for CI, cutting releases, etc.,
where it's assumed that $GOPATH is pristine.

[finishes #64960670]